aviyel/aviyel_api/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Django==4.0.2
djangorestframework==3.13.1
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
isodate==0.6.1
mypy-extensions==0.4.3
numpy==1.22.2
pandas==1.4.1
pathspec==0.9.0
platformdirs==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
